+++ cmd/heatmap/main.go
@@ -1,10 +1,12 @@
 package main
 
 import (
+	"fmt"
 	demo "github/chschmidt99/pt/pkg/demoscenes"
 	. "github/chschmidt99/pt/pkg/pt"
 	"image/png"
 	"os"
+	"strconv"
 )
 
 const (
@@ -14,20 +16,26 @@ const (
 )
 
 func main() {
-	//demoScene := demo.CornellBox(ASPECT_RATIO, FOV)
-	//demoScene := demo.Bunny(ASPECT_RATIO, FOV)
-	demoScene := demo.SanMiguel(ASPECT_RATIO, FOV)
+	//world := demo.CornellBox(ASPECT_RATIO, FOV)
+	world := demo.Bunny()
+	//world := demo.SanMiguel(ASPECT_RATIO, FOV)
+	//world := demo.Hairball(ASPECT_RATIO, FOV)
 
-	bvh := demoScene.Scene.CompilePHR(0.5, 6, 2)
-	//bvh := demoScene.Scene.CompileLBVH()
+	camera := NewDefaultCamera(ASPECT_RATIO, FOV)
+	bvh := world.Scene.CompilePHR(0.5, 6, 2)
+	renderer := NewHeatMapRenderer(bvh, camera)
 
-	renderer := NewHeatMapRenderer(bvh, demoScene.Cameras[0])
-	buff := NewBufferAspect(RESOLUTION, ASPECT_RATIO)
-	renderer.RenderToBuffer(buff)
-	img := buff.ToImage()
-	f, err := os.Create(demoScene.Name + ".png")
-	if err != nil {
-		panic(err)
+	for i, view := range world.ViewPoints {
+		camera.SetTransformation(view)
+		buff := NewPxlBufferAR(RESOLUTION, ASPECT_RATIO)
+		renderer.RenderToBuffer(buff)
+		img := buff.ToImage()
+		imageName := world.Name + " " + strconv.Itoa(i) + ".png"
+		f, err := os.Create(imageName)
+		if err != nil {
+			panic(err)
+		}
+		png.Encode(f, img)
+		fmt.Printf("Written image to " + imageName + ".png\n")
 	}
-	png.Encode(f, img)
 }