+++ cmd/image/main.go
@@ -1,36 +1,45 @@
 package main
 
 import (
+	"fmt"
 	demo "github/chschmidt99/pt/pkg/demoscenes"
 	. "github/chschmidt99/pt/pkg/pt"
 	"image/png"
 	"os"
+	"strconv"
 )
 
 const (
-	ASPECT_RATIO = 4.0 / 3
-	FOV          = 50.0
+	ASPECT_RATIO = 1
+	FOV          = 60.0
 	RESOLUTION   = 300
 )
 
 func main() {
-	//scene, camera := demo.CornellBox(ASPECT_RATIO, FOV)
-	demoScene := demo.Bunny(ASPECT_RATIO, FOV)
-	//demoScene := demo.SanMiguel(ASPECT_RATIO, FOV)
-	//demoScene := demo.Hairball(ASPECT_RATIO, FOV)
+	//world := demo.CornellBox(ASPECT_RATIO, FOV)
+	world := demo.Bunny()
+	//world := demo.SanMiguel(ASPECT_RATIO, FOV)
+	//world := demo.Hairball(ASPECT_RATIO, FOV)
 
-	bvh := demoScene.Scene.CompilePHR(0.4, 6, 2)
-	renderer := NewDefaultRenderer(bvh, demoScene.Cameras[0])
+	camera := NewDefaultCamera(ASPECT_RATIO, FOV)
+	bvh := world.Scene.CompilePHR(0.5, 6, 2)
+	renderer := NewDefaultRenderer(bvh, camera)
 	renderer.Closest = UnlitClosestHitShader
 	renderer.Miss = SkyMissShader
-	renderer.Spp = 300
+	renderer.Spp = 1
 	renderer.Verbose = true
-	buff := NewBufferAspect(RESOLUTION, ASPECT_RATIO)
-	renderer.RenderToBuffer(buff)
-	img := buff.ToImage()
-	f, err := os.Create(demoScene.Name + ".png")
-	if err != nil {
-		panic(err)
+
+	for i, view := range world.ViewPoints {
+		camera.SetTransformation(view)
+		buff := NewPxlBufferAR(RESOLUTION, ASPECT_RATIO)
+		renderer.RenderToBuffer(buff)
+		img := buff.ToImage()
+		imageName := world.Name + " " + strconv.Itoa(i) + ".png"
+		f, err := os.Create(imageName)
+		if err != nil {
+			panic(err)
+		}
+		png.Encode(f, img)
+		fmt.Printf("Written image to " + imageName + "\n")
 	}
-	png.Encode(f, img)
 }