+++ pkg/demoscenes/deer.go
@@ -9,15 +9,17 @@ func Deer(ar, fov float64) DemoScene {
 	scene := NewScene()
 	scene.Add(bunny)
 
-	camera := NewCamera(ar, fov, CameraTransformation{
-		LookFrom: NewVector3(500, 250, 500),
-		LookAt:   NewVector3(0, 110, 0),
-		Up:       NewVector3(0, 1, 0),
-	})
+	views := []CameraTransformation{
+		{
+			LookFrom: NewVector3(500, 250, 500),
+			LookAt:   NewVector3(0, 110, 0),
+			Up:       NewVector3(0, 1, 0),
+		},
+	}
 
 	return DemoScene{
-		Name:    "Deer",
-		Scene:   scene,
-		Cameras: []*Camera{camera},
+		Name:       "Deer",
+		Scene:      scene,
+		ViewPoints: views,
 	}
 }