+++ pkg/demoscenes/demoscene.go
@@ -3,7 +3,7 @@ package demoscenes
 import "github/chschmidt99/pt/pkg/pt"
 
 type DemoScene struct {
-	Name    string
-	Scene   *pt.Scene
-	Cameras []*pt.Camera
+	Name       string
+	Scene      *pt.Scene
+	ViewPoints []pt.CameraTransformation
 }