+++ pkg/demoscenes/sanmiguel.go
@@ -9,16 +9,16 @@ func SanMiguel(ar, fov float64) DemoScene {
 	root := NewSceneNode(NewMesh(geometry, whiteMat))
 	scene := NewScene()
 	scene.Add(root)
-
-	camera := NewCamera(ar, fov, CameraTransformation{
-		LookFrom: NewVector3(22, 2, 7),
-		LookAt:   NewVector3(15, 3, 1.5),
-		Up:       NewVector3(0, 1, 0),
-	})
-
+	views := []CameraTransformation{
+		{
+			LookFrom: NewVector3(22, 2, 7),
+			LookAt:   NewVector3(15, 3, 1.5),
+			Up:       NewVector3(0, 1, 0),
+		},
+	}
 	return DemoScene{
-		Name:    "San Miguel",
-		Scene:   scene,
-		Cameras: []*Camera{camera},
+		Name:       "San Miguel",
+		Scene:      scene,
+		ViewPoints: views,
 	}
 }