+++ pkg/pt/buffer.go
@@ -2,8 +2,6 @@ package pt
 
 import (
 	"image"
-	"image/color"
-	"math"
 )
 
 type Buffer interface {
@@ -26,53 +24,88 @@ func (px *Pixel) addSample(c Color) {
 	px.color = px.color.Add(c.Sub(px.color).Div(float64(px.samples)))
 }
 
-// Convert to image color and also Gamma correct for gamma=2.0
-func (px Pixel) GoColor() color.Color {
-	r := math.Sqrt(px.color.X)
-	g := math.Sqrt(px.color.Y)
-	b := math.Sqrt(px.color.Z)
-	return color.RGBA{R: uint8(Clamp(r, 0.0, 1.0) * 255), G: uint8(Clamp(g, 0.0, 1.0) * 255), B: uint8(Clamp(b, 0.0, 1.0) * 255), A: 255}
-}
-
 type PixelBuffer struct {
-	Width  int
-	Height int
-	Buff   []Pixel
+	width  int
+	height int
+	buff   []Pixel
 }
 
-func NewBufferAspect(height int, aspect float64) *PixelBuffer {
+func NewPxlBufferAR(height int, aspect float64) *PixelBuffer {
 	width := float64(height) * aspect
-	return NewBuffer(int(width), height)
+	return NewPxlBuffer(int(width), height)
 }
 
-func NewBuffer(width, height int) *PixelBuffer {
+func NewPxlBuffer(width, height int) *PixelBuffer {
 	return &PixelBuffer{
-		Width:  width,
-		Height: height,
-		Buff:   make([]Pixel, width*height),
+		width:  width,
+		height: height,
+		buff:   make([]Pixel, width*height),
 	}
 }
 
 func (b *PixelBuffer) addSample(x, y int, c Color) {
-	b.Buff[y*b.Width+x].addSample(c)
+	b.buff[y*b.width+x].addSample(c)
 }
 
 func (b *PixelBuffer) h() int {
-	return b.Height
+	return b.height
 }
 
 func (b *PixelBuffer) w() int {
-	return b.Width
+	return b.width
 }
 
 func (b *PixelBuffer) ToImage() image.Image {
 	topLeft := image.Point{0, 0}
-	bottomRight := image.Point{b.Width, b.Height}
+	bottomRight := image.Point{b.width, b.height}
 	img := image.NewRGBA(image.Rectangle{topLeft, bottomRight})
-	for i, color := range b.Buff {
-		x := i % b.Width
-		y := b.Height - (i / b.Width)
-		img.Set(x, y, color.GoColor())
+	for i, px := range b.buff {
+		x := i % b.width
+		y := b.height - (i / b.width)
+		img.Set(x, y, px.color.goColor())
+	}
+	return img
+}
+
+type FrameBuffer struct {
+	width  int
+	height int
+	buff   []Color
+}
+
+func NewFrameBufferAR(height int, aspect float64) *FrameBuffer {
+	width := float64(height) * aspect
+	return NewFrameBuffer(int(width), height)
+}
+
+func NewFrameBuffer(width, height int) *FrameBuffer {
+	return &FrameBuffer{
+		width:  width,
+		height: height,
+		buff:   make([]Color, width*height),
+	}
+}
+
+func (b *FrameBuffer) addSample(x, y int, c Color) {
+	b.buff[y*b.width+x] = c
+}
+
+func (b *FrameBuffer) w() int {
+	return b.width
+}
+
+func (b *FrameBuffer) h() int {
+	return b.height
+}
+
+func (b *FrameBuffer) ToImage() image.Image {
+	topLeft := image.Point{0, 0}
+	bottomRight := image.Point{b.width, b.height}
+	img := image.NewRGBA(image.Rectangle{topLeft, bottomRight})
+	for i, color := range b.buff {
+		x := i % b.width
+		y := b.height - (i / b.width)
+		img.Set(x, y, color.goColor())
 	}
 	return img
 }