+++ CHANGELOG.md
@@ -5,6 +5,11 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
 ## [Unreleased]
+### Added 
+- Frame Buffer 
+
+### Changed 
+- Camera orientation can now be set on existing cameras
 
 ## [0.0.4] - 2021-09-17
 ### Added